+++ frontend/src/app/components/containers/cases/case-activities/case-activity-card.tsx
